strands-demo/demo_doc_mcp/requirements.txt
fastapi
uvicorn
boto3
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
marked